requirements.txt
certifi
charset-normalizer
idna
pip
python-dotenv
requests
setuptools
urllib3
xmltodict
pip-system-certs
crontab